feat: Add Max Power Available (MPA) calculation
This commit introduces the calculation for Max Power Available (MPA).

Changes include:
*   Adding `calculateMpa()` and a corresponding `maxPower` configuration variable.
*   Refactoring internal variables from `Int` and `Long` to `Double` for improved precision in calculations.
*   Updating tests to reflect the new data types and adding new tests for MPA calculation.

+++ app/src/main/kotlin/com/currand60/wprimebalance/data/WPrimeDataSource.kt
@@ -56,7 +56,7 @@ class WPrimeDataSource(
                 when (it) {
                     is StreamState.Streaming -> {
                         val power = it.dataPoint.singleValue?.toInt() ?: 0
-                        val wPrimeBal = calculator.calculateWPrimeBalance(power,
+                        val wPrimeBal = calculator.calculateWPrimeBalance(power.toDouble(),
                             System.currentTimeMillis())
                         Timber.d("Updating W' Prime with wPrimeBal: $wPrimeBal")
                         emitter.onNext(
